Preparing for release
* Added some extra docs to Akka impl
* Refactored zerodep implementation to use more efficient map based
  building, and extracted more classes out of BuiltGraph.
